# crates/meta-whatsapp-server/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "meta-whatsapp-server",
    "description": "The meta-whatsapp-rs HTTP service: WhatsApp Business Platform over REST, for apps not written in Rust. Errors: every error answers `ErrorBody`; branch on `error.code` (`ErrorCode`), and resend only when `may_have_been_sent` is false.",
    "contact": {
      "name": "Vymalo"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/vault/rotate": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Re-encrypt every bound WABA's token (and its credit ledger) under the\nactive vault key (`WA_VAULT_KEY`), walking the service's bindings (the\nvault cannot list its records). The old key can go once `failed` is\nempty. Also `meta-whatsapp-server vault rotate`.",
        "description": "The walk is a request like any other, cut at the 55 s request\ndeadline: one too large to finish in time answers `504 timeout`, part\nof the records rotated. It is idempotent: repeat it (records already\nunder the active key are read, not rewritten), or run\n`meta-whatsapp-server vault rotate`, which has no deadline.",
        "operationId": "rotate_vault",
        "responses": {
          "200": {
            "description": "Walked every WABA: how many records were re-encrypted, and which failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultRotation"
                }
              }
            }
          },
          "401": {
            "description": "No valid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "`timeout`: the walk outlived the 55 s request deadline, part of the records rotated; repeat it, or run `meta-whatsapp-server vault rotate`, which has no deadline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error: branch on `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/v1/admin/wabas/{waba_id}/binding": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Unbind a WABA (decision D4: the admin unbind that lets another tenant\nconnect it), also the way to free a WABA whose token no longer works:\nwith its stored token, if usable, unsubscribe the app, at best; then\ndelete the token and the bindings of the WABA and its numbers.",
        "operationId": "unbind_waba",
        "parameters": [
          {
            "name": "waba_id",
            "in": "path",
            "description": "WhatsApp Business Account id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unbound, its token deleted (the app unsubscribed when the token allowed it)"
          },
          "401": {
            "description": "No valid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "`not_found`: not bound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "default": {
            "description": "Any other error: branch on `error.code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error answer.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject",
            "description": "The error."
          }
        }
      },
      "ErrorCode": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/KnownErrorCode"
          },
          {
            "type": "string"
          }
        ],
        "description": "Stable error code: a KnownErrorCode today. Codes only grow within v1: handle an unknown one by its HTTP status class."
      },
      "ErrorObject": {
        "type": "object",
        "description": "An error: its stable code, the service's sentence for it, whether to\nretry, whether the request may have taken effect, and Meta's error when\nMeta answered with one.",
        "required": [
          "code",
          "message",
          "retryable",
          "may_have_been_sent",
          "field",
          "step",
          "resumable",
          "graph",
          "request_id"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable error code (one of the `ErrorCode` values)."
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request field at fault, for `invalid_request`."
          },
          "graph": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GraphErrorInfo",
                "description": "Meta's error, when Meta answered with one."
              },
              {
                "type": "null"
              }
            ]
          },
          "may_have_been_sent": {
            "type": "boolean",
            "description": "Whether the request may have taken effect (a message may have been\nsent): if `true`, reconcile before repeating it."
          },
          "message": {
            "type": "string",
            "description": "The service's sentence for the code."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request's id (`X-Request-Id`)."
          },
          "resumable": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the stopped operation can be resumed."
          },
          "retryable": {
            "type": "boolean",
            "description": "Whether repeating the same request later may succeed."
          },
          "step": {
            "type": [
              "string",
              "null"
            ],
            "description": "The step a multi-step operation stopped at."
          }
        }
      },
...
      "GraphErrorInfo": {
        "type": "object",
        "description": "Meta's error, as the body carries it: the code and trace id, and, on\nthe routes that opt in, `error_data.details`. Never Meta's `message` or\ntitles.",
        "required": [
          "code",
          "subcode",
          "fbtrace_id",
          "details"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64",
            "description": "Meta's error code, e.g. `131047`."
          },
          "details": {
            "type": [
              "string",
              "null"
            ],
            "description": "Meta's own text (`error_data.details`), on the routes that keep it:\nnot the service's, and not a stable value to branch on."
          },
          "fbtrace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Trace id for Meta's support."
          },
          "subcode": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Meta's subcode, when it sent one (deprecated by Meta since v16)."
          }
        }
      },
...
      "KnownErrorCode": {
        "type": "string",
        "description": "The error codes this version answers (a client can switch on them and let any other fall to its default).",
        "enum": [
          "unauthenticated",
          "method_not_allowed",
          "invalid_request",
          "invalid_parameter",
          "unsupported_message_type",
          "recipient_not_supported",
          "undeliverable",
          "template_parameter_mismatch",
          "template_not_found",
          "template_text_too_long",
          "template_policy_violation",
          "template_rejected",
          "idempotency_key_reused",
          "customer_service_window_closed",
          "marketing_opted_out",
          "blocked_by_business",
          "experiment_holdout",
          "template_paused",
          "template_disabled",
          "template_syncing",
          "template_unavailable",
          "template_limit_reached",
          "flow_unavailable",
          "registration",
          "two_step_verification",
          "sync_not_allowed",
          "duplicate_onboarding",
          "number_not_connected",
          "reconnect_required",
          "waba_owned_by_another_tenant",
          "tenant_exists",
          "idempotency_in_progress",
          "outcome_unknown",
          "permission",
          "account_restricted",
          "country_restricted",
          "payment",
          "feature_not_available",
          "marketing_not_allowed",
          "forbidden",
          "tenant_suspended",
          "stale_attempt",
          "rate_limited",
          "pair_rate_limited",
          "spam_rate_limited",
          "ecosystem_engagement_limit",
          "classification_limit_reached",
          "too_many_requests",
          "too_many_streams",
          "not_found",
          "nothing_to_resume",
          "cursor_expired",
          "payload_too_large",
          "media_too_large",
          "service_unavailable",
          "unknown",
          "upstream",
          "integrity",
          "media_download_failed",
          "media_upload_failed",
          "onboarding_failed",
          "timeout",
          "storage_unavailable",
          "shutting_down",
          "internal"
        ]
      },
...
      "VaultRotation": {
        "type": "object",
        "description": "What a vault rotation did.",
        "required": [
          "wabas",
          "rotated",
          "failed"
        ],
        "properties": {
          "failed": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "WABAs whose records could not be read or rewritten (a record under\na key no longer configured, say): rotate again, or reconnect them,\nbefore dropping the old key."
          },
          "rotated": {
            "type": "integer",
            "description": "Records re-encrypted under the active key.",
            "minimum": 0
          },
          "wabas": {
            "type": "integer",
            "description": "WABAs walked.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wak_<key id>_<secret>",
        "description": "An API key: a tenant key, a platform key (with the WA-Tenant header) or an admin key (/v1/admin only)."
      }
    }
  },
...
}